Get rid of CliPluginRuntime
